Changes: - Add/Edit test cases - Change deserialized in ConsoleApp to null

+++ Zadanie2/Serializer/JsonSerializer.cs
@@ -17,19 +17,30 @@ namespace Serializer
 
         public static void Serialize(Object obj, string filePath)
         {
-            File.Delete(filePath);
-            using FileStream fs = new FileStream(filePath, FileMode.OpenOrCreate);
-            string serialized = JsonConvert.SerializeObject(obj, Formatting.Indented, JsonSerializer.JsonSettings);
-            fs.Write(Encoding.UTF8.GetBytes(serialized));
-            fs.Flush();
+            if (File.Exists(filePath))
+            {
+                File.Delete(filePath);
+            }
+            using (FileStream fs = new FileStream(filePath, FileMode.OpenOrCreate))
+            {
+                string serialized = JsonConvert.SerializeObject(obj, Formatting.Indented, JsonSerializer.JsonSettings);
+                fs.Write(Encoding.UTF8.GetBytes(serialized));
+                fs.Flush();
+            }
         }
 
         public static T Deserialize<T>(string filePath)
         {
-            using FileStream fs = new FileStream(filePath, FileMode.Open);
-            byte[] bytes = new byte[fs.Length];
-            fs.Read(bytes);
-            return JsonConvert.DeserializeObject<T>(Encoding.UTF8.GetString(bytes), JsonSerializer.JsonSettings);
+            if (File.Exists(filePath))
+            {
+                using (FileStream fs = new FileStream(filePath, FileMode.Open))
+                {
+                    byte[] bytes = new byte[fs.Length];
+                    fs.Read(bytes);
+                    return JsonConvert.DeserializeObject<T>(Encoding.UTF8.GetString(bytes), JsonSerializer.JsonSettings);
+                }
+            }
+            return default(T);
         }
 
     }
